+++ Logging.Server/LogViewer/HBaseViewer.cs
@@ -15,7 +15,7 @@ namespace Logging.Server.Viewer
             throw new NotImplementedException();
         }
 
-        public List<LogEntity> GetLogs(long start, long end, int appId, int[] level, string title, string msg, string source, int ip,List<string>tags, int limit = 100)
+        public List<LogEntity> GetLogs(long start, long end, int appId, int[] level, string title, string msg, string source, long ip,List<string>tags, int limit = 100)
         {
             return null;
         }
