+++ Logging.Server/LogViewer/ILogView.cs
@@ -8,7 +8,7 @@ namespace Logging.Server.Viewer
 {
    public interface ILogViewer
     {
-        List<LogEntity> GetLogs(long start, long end, int appId, int[] level, string title, string msg, string source, int ip,List<string>tags, int limit = 100);
+        List<LogEntity> GetLogs(long start, long end, int appId, int[] level, string title, string msg, string source, long ip,List<string>tags, int limit = 100);
 
         List<LogStatistics> GetStatistics(long start, long end,int appId);
